Convert package to ESM module.
