+++ lib/ClientMetadata.js
@@ -1,7 +1,6 @@
 /*!
- * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2020-2023 Digital Bazaar, Inc. All rights reserved.
  */
-
 import {InvalidRequest} from '@interop/oauth2-errors';
 
 class ClientMetadata {
