+++ tests/.eslintrc.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  env: {
+    mocha: true
+  }
+};
